refactor: split database + review bug fixes

+++ Domain/Entities/ProductionOrder.cs
@@ -7,8 +7,8 @@ public class ProductionOrder
     public Guid Id { get; set; }
     public string OrderNumber { get; set; } = null!;
     public Guid TargetPartId { get; set; }
-    public decimal Quantity { get; set; }
-    public decimal CompletedQuantity { get; set; }
+    public int Quantity { get; set; }
+    public int CompletedQuantity { get; set; }
     public ProductionStatus Status { get; set; }
     public DateTime PlannedStartDate { get; set; }
     public DateTime? ActualStartDate { get; set; }
